Partially finished goal tile solver
In its current state, it cannot solve goals immersed in their own color.
For instance, if the goals are surrounded in ColorNone, it cannot
find a ColorNone path between the two goals. However,
a solution could be to create a "border" on the outside of the path,
which would inherently create a ColorNone path between the two goals.